\name{harvard}
\alias{harvard}
\docType{data}
\title{A Harvard study on lung cancer gene expression}
\description{
  A Harvard study on lung cancer gene expression. Data is represented as
  an ExpressionSet.
}
\usage{data(harvard)}
\format{
}
\details{

  Annotation for the phenoData will be updated.
  
}
\source{
}
\references{
  Bhattacharjee et al., Classification of human lung carcinomas by mRNA
  expression profiling reveals distinct adenocarcinoma subclasses, PNAS
  2001, 98:13790-5.

  Parmigiani et al., A cross-study comparison of gene expression studies
  for the molecular classification of lung cancer, Clinical Cancer
  Research, 10:2922-2927, 2004.  
}
\examples{
data(harvard)
}
\keyword{datasets}
