\name{michigan}
\alias{michigan}
\docType{data}
\title{A Michigan study on lung cancer gene expression}
\description{
  A Michigan study on lung cancer gene expression. Data is represented as
  an ExpressionSet.
}
\usage{data(michigan)}
\format{
}
\details{

  Annotation for the phenoData will be updated.
  
}
\source{
}
\references{
  Beer et al., Gene expression profiles predict survival of patients
  with lung adenocarcinoma.  Nature Medicine 8(8):816-824 (2002).

  Parmigiani et al., A cross-study comparison of gene expression studies
  for the molecular classification of lung cancer, Clinical Cancer
  Research, 10:2922-2927, 2004.    
}
\examples{
data(michigan)
}
\keyword{datasets}
