\name{stanford}
\alias{stanford}
\docType{data}
\title{Public lung cancer data from the Stanford study}
\description{
  Public lung cancer data from the Stanford study represented as an ExpressionSet
}
\usage{data(stanford)}
\format{
}
\details{

  Annotation for the phenoData will be updated.
  
}
\source{

}
\references{
  Garber et al., Diversity of Gene Expression in Adenocarcinoma of the
  Lung, PNAS, 2001, 98(24):13784-9.

  Parmigiani et al., A cross-study comparison of gene expression studies
  for the molecular classification of lung cancer, Clinical Cancer
  Research, 10:2922-2927, 2004.
}
\examples{
data(stanford)
}
\keyword{datasets}
